win screen + death screen + ammo pickup
When the player dies restart screen will properly work now. They will also get the option to leave as a button

script in main checks when the amount of zombies reaches 0. The player will be taken to the win screnn

zombies have a 66% chance to drop pickable ammo. When the player walks over the ammo, 2 additional bullets are added to their ammo